Move logging handling to a separate class
